vehicle/viss/0240-unsubscribe-success.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>
[0240: end a subscription]
</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">
<link rel="stylesheet" type="text/css" href="./style-general.css">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-config.js"></script>
<script src="./vehicle-test-helper.js"></script>

<div id="test-title">
[0240: end a subscription]
</div>
<br>
<div id="test-assertion">
<b>Test assertion:</b><br>
When a client is subscribing to a data path,<br>
 client send 'UnsubscribeRequest' of the subscription to the server,<br>
 then client should receive 'UnsubscribeSuccessResponse'<br>
 and the subscription is unsubscribed.<br>
</div>
<div id="scenario">
<br>
<b>Test Scenario:</b><br>
1. Execute 'subscribe' method with a data path successfully<br>
   and receiving 'SubscriptionNotification' message continuously.<br>
   (e.g. path = 'Signal.Drivetrain.Transmission.Speed')<br>
2. Execute 'unsubscribe' method with subscriptionId of the subscription.<br>
3. Confirm client receive 'UnsubscribeSuccessResponse'.<br>
<br>
<b>Expected Result:</b><br>
- Receives 'UnsubscribeSuccessResponse' by executing 'unsubscribe' method<br>
</div>

<div id="result"></div>
<br>
<div id="log"></div>

<script>
var reqId = "reqid-0240";
var subId = null;
var path = SUBSCRIBE_STANDARD_PATH;

var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

var phase = PH_START;

var t = async_test("Unsubscribe Test: End a subscription");
var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

vehicle.onopen = function() {
  // process messages from the server
  vehicle.onmessage = function(event) {
    addLogMessage("Response: json="+event.data);
    var msg = JSON.parse(event.data);

    // subscirbe has sent
    if (phase == PH_START) {
      // nothing to do
    } else if (phase == PH_SENT_SUBSCRIBE) {
      t.step_func(function() {
        // wait for subscribe response
        if (msg.action !== "subscribe") {
          addLogMessage("PH_SENT_SUBSCRIBE: Ignore messages except for Subscribe responses.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        if (isSubscribeSuccessResponse(reqId, msg)) {
          addLogMessage("SubscribeSuccessResponse received. subscribeId = " + msg.subscriptionId);
          subId = msg.subscriptionId

          // Test successed. Send unsubscribe
          addLogMessage("<br>Send unsubscribe request");
          var unsub_request = '{"action":"unsubscribe"'
                             +' ,"subscriptionId":"'+subId+'"'
                             +' ,"requestId":"'+reqId+'"'
                             +'}';
          addLogMessage("Unsub request: json="+unsub_request);
          vehicle.send(unsub_request);

          phase = PH_SENT_UNSUBSCRIBE;
        } else if (isSubscribeErrorResponse(reqId, msg)) {
          helper_terminate_failure("SubscribeErrorResponse received. test failed.");
          phase = PH_END;
        }
      })();
    } else if (phase == PH_SENT_UNSUBSCRIBE) {
      t.step_func(function() {
        // wait for unsubscribe response
        if (msg.action !== "unsubscribe") {
          addLogMessage("Ignore messge except for Unsubscribe response.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        var res = isUnsubscribeSuccessResponse(reqId, subId, msg);
        assert_true(res, "True if UnsubscribeSuccessResponse received.");
        if (res) {
          // unsubscibe success
          phase = PH_UNSUBSCRIBED;
          addLogMessage("Test Success: UnsubscribeSuccessResponse received");
          helper_terminate_success("UnsubscribeSuccessResponse received");
        } else {
          // unsubscibe error
          addLogMessage("Test Failed: UnsubscribeErrorResponse received");
          helper_terminate_failure("UnsubscribeErrorResponse received");
        }
      })();
    } else if (phase == PH_UNSUBSCRIBED) {
      // nothing to do
    }
  }

  // in case server doesn't return response
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {
      helper_terminate_failure("Finish by timeout");
    }
  }, TIME_OUT_TIME);

  addLogMessage("Send subscribe request: path="+path);
  var sub_request = '{"action":"subscribe","path":"'+ path +'" ,"requestId":"'+reqId+'"}';
  addLogMessage("Sub request: json="+sub_request);
  vehicle.send( sub_request );
  phase = PH_SENT_SUBSCRIBE;
}
</script>
</html>
